+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## 0.1.2
+
+* Feat: notification created on Android is now `PRIORITY_LOW`
+* Refacto: `androidConfiguration` is now `androidSettings`
+
 ## 0.1.1
 
 * Feat: allow to set a custom icon for the notification on Android
+* Refacto: `AndroidBadgeConfiguration` is now `AndroidSettings`
 
 ## 0.1.0